Update zerox.py
ZeroX is another DEX that plays a key role in arbitrage and swap, striving to find the best route between specific token pairs to deliver the best possible trade. It will be one of the DEXs that plays a key role in both these areas and the swap aspect going forward.